+++ web/app/containers/checkout-shipping/constants.js
@@ -1,2 +1,3 @@
 export const SHIPPING_FORM_NAME = 'shippingForm'
 export const ADD_NEW_ADDRESS_FIELD = 'addNewAddressField'
+export const SAVED_SHIPPING_ADDRESS_FIELD = 'saved_address'
