+++ web/app/containers/checkout-shipping/partials/shipping-method.jsx
@@ -14,7 +14,7 @@ import ShippingMethodLabel from './shipping-method-label'
 
 const ShippingMethod = ({shippingMethods}) => {
     return (
-        <div>
+        <div className="t-checkout-shipping__shipping-method">
             <div className="t-checkout-shipping__title u-padding-top-lg u-padding-bottom-md">
                 <h2 className="u-h4 u-text-uppercase">Shipping Method</h2>
             </div>
